+++ stravalib/tests/integration/test_client.py
@@ -83,6 +83,7 @@ def test_update_activity(
     (
         ('file', 'tcx', {}, {'data_type': 'tcx'}, None, None),
         ('str', 'tcx', {}, {'data_type': 'tcx'}, None, None),
+        ('bytes', 'tcx', {}, {'data_type': 'tcx'}, None, None),
         ('not_supported', 'tcx', {}, {}, None, TypeError),
         ('file', 'invalid', {}, {}, None, ValueError),
         ('file', 'tcx', {'name': 'name'}, {'data_type': 'tcx', 'name': 'name'}, None, None),
@@ -135,6 +136,8 @@ def test_upload_activity(
             _call_upload(f)
         elif activity_file_type == 'str':
             _call_upload(f.read())
+        elif activity_file_type == 'bytes':
+            _call_upload(f.read().encode('utf-8'))
         else:
             _call_upload({})